fix(soql): make integration tests more robust
Use a different filename every time for every test run.

Seems like the integration test framework runs a failed test multiple times.
And in the case of language diagnostics, the second time it run for the exact same
file and input, no diagnostics check happens (I guess this is an optimization on
VSCode. That's also how it haves when testing manually).